requirements.txt
flask==2.3.2
torch==2.0.1
transformers==4.30.2
Pillow==9.4.0
python-dotenv==1.0.0
